+++ CHANGELOG.md
@@ -11,6 +11,8 @@ FEATURES:
 - **New Resource:** `netapp-ontap_fpolicy_event` ([#624](https://github.com/NetApp/terraform-provider-netapp-ontap/issues/624))
 - **New Data Source:** `netapp-ontap_network_ip_service_policy` and `netapp-ontap_network_ip_service_policies`([#621](https://github.com/NetApp/terraform-provider-netapp-ontap/issues/621))
 - **New Resource:** `netapp-ontap_network_ip_service_policy` ([#621](https://github.com/NetApp/terraform-provider-netapp-ontap/issues/621))
+- **New Data Source:** `netapp-ontap_fpolicy_external_engine` and `netapp-ontap_fpolicy_external_engines` ([#625](https://github.com/NetApp/terraform-provider-netapp-ontap/issues/625))
+- **New Resource:** `netapp-ontap_fpolicy_external_engine` ([#625](https://github.com/NetApp/terraform-provider-netapp-ontap/issues/625))
 
 ENHANCEMENTS: